+++ crates/re_renderer/examples/renderer_standalone.rs
@@ -571,7 +571,7 @@ fn main() {
     #[cfg(not(target_arch = "wasm32"))]
     {
         if std::env::var("RUST_LOG").is_err() {
-            std::env::set_var("RUST_LOG", "info,wgpu_core=off");
+            std::env::set_var("RUST_LOG", "info,wgpu_core=warn,wgpu_hal=warn");
         }
         tracing_subscriber::fmt::init();