+++ crates/rerun/src/main.rs
@@ -5,7 +5,7 @@ static GLOBAL: mimalloc::MiMalloc = mimalloc::MiMalloc;
 async fn main() -> anyhow::Result<()> {
     if std::env::var("RUST_LOG").is_err() {
         // Enable logging unless the user opts-out of it.
-        std::env::set_var("RUST_LOG", "info,wgpu_core=off");
+        std::env::set_var("RUST_LOG", "info,wgpu_core=warn,wgpu_hal=warn");
     }
     tracing_subscriber::fmt::init(); // log to stdout